refactor: add a couple of the modules to the app module
The app module was missing a few of the newer modules, so it would fail to have some imports
