+++ app/build.gradle
@@ -72,6 +72,9 @@ dependencies {
     }
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
+    implementation project(path: ':account')
+    implementation project(path: ':core')
+    implementation project(path: ':database')
     implementation project(path: ':login')
     implementation project(path: ':server')
     implementation project(path: ':kitsu')
