+++ app/src/main/java/com/chesire/malime/injection/modules/CoroutineModule.kt
@@ -1,6 +1,6 @@
 package com.chesire.malime.injection.modules
 
-import com.chesire.malime.IOContext
+import com.chesire.malime.core.IOContext
 import dagger.Module
 import dagger.Provides
 import kotlinx.coroutines.Dispatchers
@@ -10,7 +10,7 @@ import kotlin.coroutines.CoroutineContext
 @Module
 object CoroutineModule {
     @Provides
-    @IOContext
+    @com.chesire.malime.core.IOContext
     @JvmStatic
     fun providesIOContext(): CoroutineContext = Dispatchers.IO
 }
